Updated README.md

+++ src/base64_encdec.c
@@ -223,7 +223,7 @@ int32_t base64_decode(int8_t *input, size_t size, int8_t *output, size_t *capaci
 	return BASE64_ISOKAY;
 }
 
-void base64_geterrormessage(int8_t errno, int8_t *message, size_t capacity)
+void base64_geterrormessage(int32_t errno, int8_t *message, size_t capacity)
 {
 	const char messages[][64] = 
 	{
